Initial work on header file amalgamation
